+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 development:
   * prevent tux to be required
+  * add sinatra assets pipeline extension
 
 0.4.0:
   * add SQL logger to development environment